chore: rename pages folder and move error pages there
