+++ src/app/graphiql/page.tsx
@@ -1,6 +1,6 @@
 import { type ReactNode } from 'react';
 
-import GraphiQl from '@/pages/graphiql/Graphiql';
+import GraphiQl from '@/page-components/graphiql/Graphiql';
 
 export default function GraphiQlPage(): ReactNode {
   return <GraphiQl />;
