+++ src/app/rest/page.tsx
@@ -1,6 +1,6 @@
 import { type ReactNode } from 'react';
 
-import Rest from '@/pages/rest/Rest';
+import Rest from '@/page-components/rest/Rest';
 
 export default function RestPage(): ReactNode {
   return <Rest />;
