+++ src/app/sign-in/page.tsx
@@ -1,6 +1,6 @@
 import { type ReactNode } from 'react';
 
-import SignIn from '@/pages/sign-in/SignIn';
+import SignIn from '@/page-components/sign-in/SignIn';
 
 export default function SignInPage(): ReactNode {
   return <SignIn />;
