+++ src/app/sign-up/page.tsx
@@ -1,6 +1,6 @@
 import { type ReactNode } from 'react';
 
-import SignUp from '@/pages/sign-up/SignUp';
+import SignUp from '@/page-components/sign-up/SignUp';
 
 export default function SignUpPage(): ReactNode {
   return <SignUp />;
